+++ test/pipeful_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "test_helper"
-require "binding_of_caller"  # for eval mode (experimental) for using functions in local scope
+require "binding_of_caller"  # for eval mode (experimental) for piping local variables
 
 class PipefulTest < Minitest::Test
   module Pipelines
@@ -18,10 +18,10 @@ class PipefulTest < Minitest::Test
         GClassMethod >>      # == 121
         HWithBlock do |n1, n2|
           +[n1, n2] >>       # == +[121, 5]
-            IConstant >>     # == +[121, 5]
+            I_CONSTANT >>    # == +[121, 5]
             42 >>            # == +[121, 5, 42]
             JNotFunction     # == JNotFunction object containing [121, 5, 42]
-      end
+        end
     end
 
     class ASimple
@@ -95,7 +95,7 @@ class PipefulTest < Minitest::Test
       end
     end
 
-    IConstant = ->(*args) { +[*args] }
+    I_CONSTANT = ->(*args) { +[*args] }
 
     class JNotFunction
       extend Forwardable
@@ -108,18 +108,16 @@ class PipefulTest < Minitest::Test
     # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
     def self.parenthetical_args
-      +[3, 12] >>
-        InstanceMethodGrow(operator: :*) >>   # parenth. args go to initialize
-                                              #   same as ".new(operator: :*).call(3, 12)
-        ClassMethodShrink(4, operator: :/)    # or added to pipe arg(s) if .call is class method
-                                              #   same as ".call(36, 4, operator: :/)
+      +[0, 3, 12] >>
+        InstanceFunct(operator: :*) >>        # here parenth. args go to initialize
+                                              #     same as ".new(operator: :*).call(3, 12)
+        ClassFunct(4, operator: :/) >>        # they're added on to piped arg(s) if .call is class method
+                                              #     same as ".call(36, 4, operator: :/)
+        Array
     end
 
-    class InstanceMethodGrow
-      OPERATORS = %i[+ * **]
-
+    class InstanceFunct
       def initialize(operator: :+)
-        raise ArgumentError unless OPERATORS.include?(operator)
         @op = operator
       end
 
@@ -128,11 +126,8 @@ class PipefulTest < Minitest::Test
       end
     end
 
-    class ClassMethodShrink
-      OPERATORS = %i[- /]
-
+    class ClassFunct
       def self.call(x, y, operator: :-)
-        raise ArgumentError unless OPERATORS.include?(operator)
         x.send(operator, y)
       end
     end
@@ -174,25 +169,123 @@ class PipefulTest < Minitest::Test
     # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
     def self.eval_mode
-      funct_local = ->(n) { n * 2 }
+      funct_local = ->(n, mult:) { n * mult }
       @funct_instance = -> (n) { n + 1 }
 
-      Pipeful.eval_mode do
+      pipe_eval do
         5 >>
-          funct_local >>
+          funct_local(mult: 2) >>
           @funct_instance
+      end
+    end
+
+    # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+    def self.my_method_missing
+      WithMethodMissing.call
+    end
+
+    class WithMethodMissing
+      extend Pipeful
+
+      def self.call
+        +[3, 4] >>
+          InstanceFunct(operator: :*) >>
+          NonexistentFunct(1) >>
+          nonexistent_method(2) >>
+          Array
+      end
+
+      def self.method_missing(m, *args, **kwargs, &block)
+        args.first
+      end
+    end
+
+    # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+    def self.eval_beside_method_missing
+      EvalBesideMethodMissing.call
+    end
+
+    class EvalBesideMethodMissing
+      extend Pipeful
+
+      def self.call
+        multiply_by = ->(n, m) { n * m }
+
+        pipe_eval do
+          3 >>
+            multiply_by(4)
+        end
+      end
+
+      def self.method_missing(_m, *args, **_kwargs, &_block)
+        args.first
+      end
+    end
+
+    # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+    def self.eval_using_method_missing
+      EvalUsingMethodMissing.call
+    end
+
+    class EvalUsingMethodMissing
+      extend Pipeful
+
+      def self.call
+        multiply_by = ->(n, m) { n * m }
+
+        pipe_eval do
+          3 >>
+            multiply_by(4) >>
+            NonexistentFunct(100) >>
+            doesnt_exist_either(50) >>
+            multiply_by(4) >>
+            Array
+        end
+      end
+
+      def self.method_missing(_m, *args, **_kwargs, &_block)
+        args.first
+      end
+    end
+
+    # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+    def self.complex_nested_functions
+      10 >> EvalParenthNested >> Array
+    end
+
+    class EvalParenthNested
+      include Pipeful
+
+      def call(n)
+        n >>
+          InsideEvalParenth(3, operator: :+) >>
+          DoesntExist(300)
+      end
+
+      class InsideEvalParenth
+        def self.call(a, b, operator:)
+          a.send(operator, b)
+        end
+      end
+
+      def method_missing(_m, *args, **_kwargs, &_block)
+        args.first
       end
     end
   end
 
-  # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+  # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
   def test_basic
-    assert_equal [121, 5, 42], Pipelines.basic.to_a
+   assert_equal [121, 5, 42], Pipelines.basic.to_a
   end
 
   def test_parenthetical_args
-    assert_equal 9, Pipelines.parenthetical_args
+    assert_equal [0, 9], Pipelines.parenthetical_args
   end
 
   def test_module_notation
@@ -201,6 +294,22 @@ class PipefulTest < Minitest::Test
 
   def test_eval_mode
     assert_equal 11, Pipelines.eval_mode
+  end
+
+  def test_method_missing_works
+    assert_equal [12, 1, 2], Pipelines.my_method_missing
+  end
+
+  def test_eval_beside_method_missing_works
+    assert_equal 12, Pipelines.eval_beside_method_missing
+  end
+
+  def test_eval_using_method_missing_also_works
+    assert_equal [12, 100, 200], Pipelines.eval_using_method_missing
+  end
+
+  def test_complex_nested_functions
+    assert_equal [13, 300], Pipelines.complex_nested_functions
   end
 
   def test_that_it_has_a_version_number